WIP: moves Negotiation Entity into its package
