+++ src/main/java/eu/bbmri_eric/negotiator/configuration/state_machine/negotiation/PersistStateChangeListener.java
@@ -1,6 +1,6 @@
 package eu.bbmri_eric.negotiator.configuration.state_machine.negotiation;
 
-import eu.bbmri_eric.negotiator.database.model.Negotiation;
+import eu.bbmri_eric.negotiator.database.model.negotiation.Negotiation;
 import eu.bbmri_eric.negotiator.database.repository.NegotiationRepository;
 import eu.bbmri_eric.negotiator.database.repository.NotificationRepository;
 import eu.bbmri_eric.negotiator.database.repository.PersonRepository;
