+++ src/main/java/eu/bbmri_eric/negotiator/database/model/attachment/AttachmentProjection.java
@@ -1,0 +1,5 @@
+package eu.bbmri_eric.negotiator.database.model.attachment;
+
+public interface AttachmentProjection extends MetadataAttachmentProjection {
+  byte[] getPayload();
+}
