+++ src/main/java/eu/bbmri_eric/negotiator/database/model/attachment/MetadataAttachmentProjection.java
@@ -1,8 +1,11 @@
-package eu.bbmri_eric.negotiator.database.model.attachments;
+package eu.bbmri_eric.negotiator.database.model.attachment;
 
 public interface MetadataAttachmentProjection extends BaseAttachmentProjection {
   String getId();
+
   String getName();
+
   String getSize();
+
   String getContentType();
 }
