--- src/main/java/eu/bbmri_eric/negotiator/database/model/attachments/AttachmentProjection.java
@@ -1,5 +1,0 @@
-package eu.bbmri_eric.negotiator.database.model.attachments;
-
-public interface AttachmentProjection extends MetadataAttachmentProjection {
-    byte[] getPayload();
-}
