+++ src/main/java/eu/bbmri_eric/negotiator/database/repository/NegotiationRepository.java
@@ -2,7 +2,7 @@ package eu.bbmri_eric.negotiator.database.repository;
 
 import eu.bbmri_eric.negotiator.configuration.state_machine.negotiation.NegotiationState;
 import eu.bbmri_eric.negotiator.configuration.state_machine.resource.NegotiationResourceState;
-import eu.bbmri_eric.negotiator.database.model.Negotiation;
+import eu.bbmri_eric.negotiator.database.model.negotiation.Negotiation;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
