+++ src/main/java/eu/bbmri_eric/negotiator/service/DBAttachmentService.java
@@ -1,13 +1,13 @@
 package eu.bbmri_eric.negotiator.service;
 
 import eu.bbmri_eric.negotiator.configuration.security.auth.NegotiatorUserDetailsService;
-import eu.bbmri_eric.negotiator.database.model.Negotiation;
 import eu.bbmri_eric.negotiator.database.model.Organization;
 import eu.bbmri_eric.negotiator.database.model.Person;
 import eu.bbmri_eric.negotiator.database.model.Resource;
-import eu.bbmri_eric.negotiator.database.model.attachments.Attachment;
-import eu.bbmri_eric.negotiator.database.model.attachments.BaseAttachmentProjection;
-import eu.bbmri_eric.negotiator.database.model.attachments.MetadataAttachmentProjection;
+import eu.bbmri_eric.negotiator.database.model.attachment.Attachment;
+import eu.bbmri_eric.negotiator.database.model.attachment.BaseAttachmentProjection;
+import eu.bbmri_eric.negotiator.database.model.attachment.MetadataAttachmentProjection;
+import eu.bbmri_eric.negotiator.database.model.negotiation.Negotiation;
 import eu.bbmri_eric.negotiator.database.repository.AttachmentRepository;
 import eu.bbmri_eric.negotiator.database.repository.NegotiationRepository;
 import eu.bbmri_eric.negotiator.database.repository.OrganizationRepository;
@@ -147,7 +147,8 @@ public class DBAttachmentService implements AttachmentService {
   @Override
   @Transactional
   public List<AttachmentMetadataDTO> findByNegotiation(String id) {
-    List<MetadataAttachmentProjection> attachments = attachmentRepository.findMetadataByNegotiationId(id);
+    List<MetadataAttachmentProjection> attachments =
+        attachmentRepository.findMetadataByNegotiationId(id);
     return attachments.stream()
         .filter(this::isAuthorizedForAttachment)
         .map((attachment) -> modelMapper.map(attachment, AttachmentMetadataDTO.class))
