+++ src/test/java/eu/bbmri_eric/negotiator/unit/model/NotificationTest.java
@@ -2,9 +2,9 @@ package eu.bbmri_eric.negotiator.unit.model;
 
 import static org.junit.jupiter.api.Assertions.assertInstanceOf;
 
-import eu.bbmri_eric.negotiator.database.model.Negotiation;
 import eu.bbmri_eric.negotiator.database.model.Notification;
 import eu.bbmri_eric.negotiator.database.model.Person;
+import eu.bbmri_eric.negotiator.database.model.negotiation.Negotiation;
 import org.junit.jupiter.api.Test;
 
 public class NotificationTest {
